Module crosvm_direct::crosvm

The root level module that includes the config and aggregate of the submodules for running said configs.

Modules

argument

Handles argument parsing.

gdb

platform

plugin

Structs

BalloonCommand

Set balloon size of the crosvm instance to SIZE bytes

BalloonStatsCommand

Prints virtio balloon statistics for a VM_SOCKET

BatteryCommand

Modify battery

BindMount

A bind mount for directories in the plugin process.

Config

Aggregate of all configurable options for a running VM.

CreateCompositeCommand

Create a new composite disk image file

CreateQcow2Command

Create Qcow2 image given path and size

CrosvmCmdlineArgs

crosvm

DevicesCommand

Start a device process

DirectIoOption

Direct IO forwarding options

DiskCommand

Manage attached virtual disk devices

FileBackedMappingParameters

GidMap

A mapping of linux group IDs for the plugin process.

GpeCommand

Injects a general-purpose event into the crosvm instance

HostPcieRootPortParameters

JailConfig

MakeRTCommand

Enables real-time vcpu priority for crosvm instances started with --delay-rt

PowerbtnCommand

Triggers a power button event in the crosvm instance

ResizeDiskSubcommand

resize disk

ResumeCommand

Resumes the crosvm instance

RunCommand

RunCommandCatchAllInternal

Start a new crosvm instance

SharedDir

SleepCommand

Triggers a sleep button event in the crosvm instance

StopCommand

Stops crosvm instances via their control sockets

SuspendCommand

Suspends the crosvm instance

TouchDeviceOption

UsbAttachCommand

Attach usb device

UsbCommand

Manage attached virtual USB devices.

UsbDetachCommand

Detach usb device

UsbListCommand

Detach usb device

VersionCommand

Show package version.

VfioAddSubCommand

ADD

VfioCommand

VFIO device structure for creating a new instance based on command line options.

VfioCrosvmCommand

add/remove host vfio pci device into guest

VfioRemoveSubCommand

REMOVE

VhostUserFsOption

VhostUserOption

VhostUserWlOption

VvuOption

Options for virtio-vhost-user proxy device.

Enums

Command

CrossPlatformDevicesCommands

Cross-platform Devices

DevicesSubcommand

DiskSubcommand

Executable

Indicates the location and kind of executable kernel for a VM.

SharedDirKind

UsbSubCommand

VfioSubCommand

VfioType

Vfio device type, recognized based on command line option.

Constants

DEFAULT_TOUCH_DEVICE_HEIGHT

DEFAULT_TOUCH_DEVICE_WIDTH

Statics

KVM_PATH

SECCOMP_POLICY_DIR

VHOST_NET_PATH

Functions

parse_bus_id_addr