Module crosvm_direct::crosvm::config

Structs

BindMount

A bind mount for directories in the plugin process.

Config

Aggregate of all configurable options for a running VM.

DirectIoOption

Direct IO forwarding options

FileBackedMappingParameters

GidMap

A mapping of linux group IDs for the plugin process.

GpuDisplayParametersBuilder

HostPcieRootPortParameters

JailConfig

SharedDir

TouchDeviceOption

VfioCommand

VFIO device structure for creating a new instance based on command line options.

VhostUserFsOption

VhostUserOption

VhostUserWlOption

VvuOption

Options for virtio-vhost-user proxy device.

Enums

Executable

Indicates the location and kind of executable kernel for a VM.

SharedDirKind

VfioType

Vfio device type, recognized based on command line option.

Constants

DEFAULT_TOUCH_DEVICE_HEIGHT

DEFAULT_TOUCH_DEVICE_WIDTH

MAX_PCIE_ECAM_SIZE

MB_ALIGNED

ONE_MB

Statics

KVM_PATH

SECCOMP_POLICY_DIR

VHOST_NET_PATH

Functions

executable_is_plugin

from_key_values

invalid_value_err

numbered_disk_option

parse_ac97_options

parse_battery_options

parse_bus_id_addr

parse_cpu_affinity

Parse a list of guest to host CPU mappings.

parse_cpu_capacity

parse_cpu_set

Parse a comma-separated list of CPU numbers and ranges and convert it to a Vec of CPU numbers.

parse_direct_io_options

parse_file_backed_mapping

parse_gpu_display_options

parse_gpu_options

parse_memory_region

parse_mmio_address_range

parse_pcie_root_port_params

parse_plugin_mount_option

parse_pstore

parse_serial_options

parse_stub_pci_parameters

parse_userspace_msr_options

parse_vfio

parse_vfio_platform

parse_video_options

validate_config