Module crosvm_direct::crosvm::platform::device_helpers

Enums

TaggedControlTube

Traits

IntoUnixStream

Functions

add_bind_mounts

create_9p_device

create_balloon_device

create_block_device

create_console_device

create_fs_device

create_iommu_device

create_keyboard_device

create_mouse_device

create_multi_touch_device

create_net_device

Generic method for creating a network device. create_device is a closure that takes the virtio features and number of queue pairs as parameters, and is responsible for creating the device itself.

create_net_device_from_config

Returns a network device created from a new TAP interface configured with host_ip, netmask, and mac_address.

create_pmem_device

create_rng_device

create_single_touch_device

create_software_tpm_device

create_sound_device

create_switches_device

create_tap_net_device_from_fd

Returns a network device from a file descriptor to a configured TAP interface.

create_tap_net_device_from_name

Returns a network device created by opening the persistent, configured TAP interface tap_name.

create_trackpad_device

create_vfio_device

create_vfio_platform_device

create_vhost_user_block_device

create_vhost_user_console_device

create_vhost_user_fs_device

create_vhost_user_mac80211_hwsim_device

create_vhost_user_net_device

create_vhost_user_snd_device

create_vhost_user_vsock_device

create_vhost_user_wl_device

create_vhost_vsock_device

create_video_device

create_vinput_device

create_vvu_proxy_device

create_wayland_device

register_video_device

setup_virtio_access_platform

Setup for devices with virtio-iommu

Type Definitions

DeviceResult