Function crosvm_direct::crosvm::sys::unix::jail_and_start_vu_device

fn jail_and_start_vu_device<T: VirtioDeviceBuilder>(
    jail_config: &Option<JailConfig>, 
    params: &T, 
    vhost: &str, 
    name: &str
) -> Result<(pid_t, Option<Box<dyn Any>>)>

Start and jail a vhost-user device according to its configuration and a vhost listener string.

The jailing business is nasty and potentially unsafe if done from the wrong context - do not call outside of start_devices!

Returns the pid of the jailed device process.