Module crosvm_direct::crosvm::cmdline

Structs

BalloonCommand

Set balloon size of the crosvm instance to SIZE bytes

BalloonStatsCommand

Prints virtio balloon statistics for a VM_SOCKET

BatteryCommand

Modify battery

CreateCompositeCommand

Create a new composite disk image file

CreateQcow2Command

Create Qcow2 image given path and size

CrosvmCmdlineArgs

crosvm

DeviceCommand

Start a device process

DiskCommand

Manage attached virtual disk devices

GpeCommand

Injects a general-purpose event into the crosvm instance

MakeRTCommand

Enables real-time vcpu priority for crosvm instances started with --delay-rt

PowerbtnCommand

Triggers a power button event in the crosvm instance

ResizeDiskSubcommand

resize disk

ResumeCommand

Resumes the crosvm instance

RunCommand

Start a new crosvm instance

SleepCommand

Triggers a sleep button event in the crosvm instance

StopCommand

Stops crosvm instances via their control sockets

SuspendCommand

Suspends the crosvm instance

UsbAttachCommand

Attach usb device

UsbCommand

Manage attached virtual USB devices.

UsbDetachCommand

Detach usb device

UsbListCommand

Detach usb device

VersionCommand

Show package version.

VfioAddSubCommand

ADD

VfioCrosvmCommand

add/remove host vfio pci device into guest

VfioRemoveSubCommand

REMOVE

Enums

Command

CrossPlatformDevicesCommands

Cross-platform Devices

DeviceSubcommand

DiskSubcommand

UsbSubCommand

VfioSubCommand