Struct crosvm_direct::crosvm::plugin::vcpu::PluginVcpu

pub struct PluginVcpu {
    shared_vcpu_state: Arc<RwLock<SharedVcpuState>>,
    per_vcpu_state: Arc<Mutex<PerVcpuState>>,
    read_pipe: File,
    write_pipe: File,
    wait_reason: Cell<Option<VcpuResponse_Wait>>,
    request_buffer: RefCell<Vec<u8>>,
    response_buffer: RefCell<Vec<u8>>,
}

State object for a VCPU’s connection with the plugin process.

This is used by a VCPU thread to allow the plugin process to handle vmexits. Each method may block indefinitely while the plugin process is handling requests. In order to cleanly shutdown during these blocking calls, the connection socket should be shutdown. This will end the blocking calls,

Fields
shared_vcpu_state: Arc<RwLock<SharedVcpuState>>per_vcpu_state: Arc<Mutex<PerVcpuState>>read_pipe: Filewrite_pipe: Filewait_reason: Cell<Option<VcpuResponse_Wait>>request_buffer: RefCell<Vec<u8>>response_buffer: RefCell<Vec<u8>>
Implementations

impl PluginVcpu

pub fn new(
shared_vcpu_state: Arc<RwLock<SharedVcpuState>>,
per_vcpu_state: Arc<Mutex<PerVcpuState>>,
read_pipe: File,
write_pipe: File
) -> PluginVcpu

Creates the plugin state and connection container for a VCPU thread.

pub fn init(&self, vcpu: &Vcpu) -> SysResult<()>

Tells the plugin process to initialize this VCPU.

This should be called for each VCPU before the first run of any of the VCPUs in the VM.

pub fn pre_run(&self, vcpu: &Vcpu) -> SysResult<()>

The VCPU thread should call this before rerunning a VM in order to handle pending requests to this VCPU.

fn process(
&self,
io_space: IoSpace,
addr: u64,
data: VcpuRunData<'_>,
vcpu: &Vcpu
) -> bool

pub fn io_read(&self, addr: u64, data: &mut [u8], vcpu: &Vcpu) -> bool

Has the plugin process handle a IO port read.

pub fn io_write(&self, addr: u64, data: &[u8], vcpu: &Vcpu) -> bool

Has the plugin process handle a IO port write.

pub fn mmio_read(&self, addr: u64, data: &mut [u8], vcpu: &Vcpu) -> bool

Has the plugin process handle a MMIO read.

pub fn mmio_write(&self, addr: u64, data: &[u8], vcpu: &Vcpu) -> bool

Has the plugin process handle a MMIO write.

pub fn hyperv_call(
&self,
input: u64,
params: [u64; 2],
data: &mut [u8],
vcpu: &Vcpu
) -> bool

Has the plugin process handle a hyper-v call.

pub fn hyperv_synic(
&self,
msr: u32,
control: u64,
evt_page: u64,
msg_page: u64,
vcpu: &Vcpu
) -> bool

Has the plugin process handle a synic config change.

fn handle_request(&self, vcpu: &Vcpu) -> SysResult<Option<Vec<u8>>>

fn handle_until_resume(&self, vcpu: &Vcpu) -> SysResult<Vec<u8>>